refactor: LottoValidator을 global의 Validator로 통합
Store과 Lotto의 유효성 검사 중복 내용을 피하기 위해 하나로 통합
